Upgrade[code]: file schema
