+++ makefile
@@ -1,11 +1,11 @@
 CC=clang -std=gnu11 
-base=main.c routine.c reuse.c routine_helper.c
-freebsd: $(base) routine_freebsd.s
+base=main_t.c src/routine.c src/reuse.c src/routine_helper.c
+freebsd: $(base) src/routine_freebsd.s
 	$(CC) $^ -o $@ 
 
-linux: $(base) routine_linux.s
+linux: $(base) src/routine_linux.s
 	$(CC) $^ -o $@
 
 .PHONY:clean
 clean:
-	rm *.o $(base)
+	rm *.o freebsd linux
