+++ src/reuse.c
@@ -1,5 +1,2 @@
-typedef struct _reuse
-{
-    struct _reuse *prev;
-    struct _reuse *next;
-} reuse, *preuse;
+#include "reuse.h"
+
