+++ src/routine.c
@@ -2,10 +2,12 @@
 
 proutine head;
 proutine tail;
-
 pcollect coll;
+proutine ROUTINE_NR[1024];
+#define SET_RET_U_ROUTINE \
+    proutine r = create_routine0(p); \
+    const uroutine u = {r->rid, &(r->rax), &(r->status)}; \
 
-proutine NR[STACK_LEN];
 
 data_p stack_from_collection() {
     data_p s = coll->stack;
@@ -17,50 +19,47 @@ data_p acquire_stack0(int len) {
     return coll == NULL ? malloc(len * sizeof(data_t)) : stack_from_collection();
 }
 
-proutine create_routine0(any p) {
-    if (NR[0]==NULL){
-       NR[0] = create_current_routine();
-    }
+void set_rid(proutine r) {
+    static int curr_pid = 0;
+    ROUTINE_NR[r->rid = ++curr_pid] = r;
+}
+
+proutine init_routine() {
     proutine r = malloc(sizeof(routine));
-    create_rid(r);
-    init_stack(&r, acquire_stack0(STACK_LEN), STACK_LEN, p);
-    //insert
-    tail->next = r;
-    tail = r;
+    memset(r, 0, sizeof(routine));
+    return r;
+}
+
+proutine create_routine0(any p) {
+    ROUTINE_NR[0] = create_current_routine();
+    proutine r = init_routine();
+    set_rid(r);
+    init_stack(r, acquire_stack0(STACK_LEN), STACK_LEN, p);
+    insert(&tail, r);
     return r;
 }
 
 uroutine create_routine_with_params(any p, int num, ...) {
-    proutine r = create_routine0(p);
-    const uroutine u = {r->rid,&(r->rax), &(r->status)};
-    va_list valist;
-    va_start(valist, num);
+    SET_RET_U_ROUTINE
+    va_list p_list;
+    va_start(p_list, num);
     data_p w = &(r->rdi);
-    for (int i = 0; i < num; i++) {
-        (*w) = va_arg(valist,data_t);
-        w++;
+    for (int i = 0; i < num; i++, w++) {
+        (*w) = va_arg(p_list, data_t);
     }
-    va_end(valist);
+    va_end(p_list);
     return u;
 }
 
 uroutine create_routine(any p) {
-    proutine r = create_routine0(p);
-    const uroutine u = {r->rid ,&(r->rax), &(r->status)};
+    SET_RET_U_ROUTINE
     return u;
 }
 
 proutine create_current_routine() {
-    proutine r = malloc(sizeof(routine));
-    tail = head = r;
-    return r;
+    return ROUTINE_NR[0] != NULL ? ROUTINE_NR[0] : (tail = head = init_routine());
 }
 
 void remove_from_bitmap(rid_t rid) {
-    NR[rid] = NULL;
-}
-
-void create_rid(proutine r){
-    static rid_t curr = 0;
-    NR[r->rid=curr++]=r;
+    ROUTINE_NR[rid] = NULL;
 }
