+++ src/routine_helper.c
@@ -27,12 +27,12 @@ void exchange_c(proutine *h, proutine *t) {
     (*h) = n;
 }
 
-void init_stack(proutine *r, data_p stack, int len, any p) {
-    (*r)->stack = stack;
+void init_stack(proutine r, data_p stack, int len, any p) {
+    r->stack = stack;
     stack[len - 1] = (code_t) stop_routine;
     stack[len - 2] = (code_t) p;
-    (*r)->esp = (code_t) ((*r)->stack + STACK_LEN - 2);
-    (*r)->ebp = (code_t) ((*r)->stack + STACK_LEN - 1);
+    r->esp = (code_t) (stack + STACK_LEN - 2);
+    r->ebp = (code_t) (stack + STACK_LEN - 1);
 }
 
 void insert(proutine *n, proutine r) {
